requirements.txt
six
Faker
pillow
markdown
python-dateutil
pytz
DateTime
folium

django
django-environ
django-filter
django-resized
python-decouple
django-crispy-forms

crispy-bootstrap5
django-browser-reload
django-phonenumber-field[phonenumbers]
stripe
django-ckeditor-5
django-jazzmin
folium
stripe